Update console-based rock-paper-scissors game to use UI elements.
Added buttons for the user to select which item to
play in a given round. The choice is highlighted
in yellow, and a simple text box shows the outcome
of the round. A rudimentarty scoreboard keeps
track of the scores, and when a player reaches 5,
that player is declared the winner.